requirements.txt
ops ~= 2.17
PyYAML ~= 6.0
